backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-django==4.7.0
setuptools>=65.0.0
tzdata==2023.3 